fix: note about the termination of Node 6 support
