+++ package.json
@@ -107,6 +107,7 @@
     "@types/lodash": "^4.14.136",
     "@types/needle": "^2.0.4",
     "@types/node": "^6.14.4",
+    "@types/semver": "^6.2.0",
     "@types/sinon": "^7.5.0",
     "@typescript-eslint/eslint-plugin": "^2.0.0",
     "@typescript-eslint/parser": "^2.0.0",
